fix: common file with provider and tokens definitions to prevent circular dependency and tgz in angular-bindings
