+++ packages/angular-bindings/src/statsig.module.ts
@@ -1,30 +1,12 @@
-import { InjectionToken, NgModule } from '@angular/core';
-
-import {
-  FeatureGateEvaluationOptions,
-  StatsigClient,
-  StatsigOptions,
-  StatsigUser,
-} from '@statsig/js-client';
+import { NgModule } from '@angular/core';
 
 import { CheckGateDirective } from './checkGate.directive';
 
-type WithClient<T extends StatsigClient> = { client: T };
-type WithConfiguration = {
-  sdkKey: string;
-  user: StatsigUser;
-  options?: StatsigOptions;
-};
-
-export type StatsigInitConfig<T extends StatsigClient> =
-  | WithClient<T>
-  | WithConfiguration;
-
-export const STATSIG_INIT_CONFIG = new InjectionToken<
-  StatsigInitConfig<StatsigClient>
->('StatsigProvider');
-
-export type FeatureGateOptions = FeatureGateEvaluationOptions;
+export {
+  FeatureGateOptions,
+  STATSIG_INIT_CONFIG,
+  StatsigInitConfig,
+} from './statsig.providers';
 
 @NgModule({
   declarations: [CheckGateDirective],
